WordPressAuthenticatorTests: Splitting LoginFields Validation Tests
