+++ WordPress/WordPressTest/WordPressAuthenticatorTests.swift
@@ -48,25 +48,6 @@ class WordPressAuthenticatorTests: XCTestCase {
     }
 
 
-    func testValidateFieldsPopulatedForSignin() {
-        let loginFields = LoginFields()
-        loginFields.meta.userIsDotCom = true
-
-        XCTAssert(!WordPressAuthenticator.validateFieldsPopulatedForSignin(loginFields), "Empty fields should not validate.")
-
-        loginFields.username = "user"
-        XCTAssert(!WordPressAuthenticator.validateFieldsPopulatedForSignin(loginFields), "Should not validate with just a username")
-
-        loginFields.password = "password"
-        XCTAssert(WordPressAuthenticator.validateFieldsPopulatedForSignin(loginFields), "should validate wpcom with username and password.")
-
-        loginFields.meta.userIsDotCom = false
-        XCTAssert(!WordPressAuthenticator.validateFieldsPopulatedForSignin(loginFields), "should not validate self-hosted with just username and password.")
-
-        loginFields.siteAddress = "example.com"
-        XCTAssert(WordPressAuthenticator.validateFieldsPopulatedForSignin(loginFields), "should validate self-hosted with username, password, and site.")
-    }
-
     func testExtractUsernameFrom() {
         let plainUsername = "auser"
         XCTAssertEqual(plainUsername, WordPressAuthenticator.extractUsername(from: plainUsername))
@@ -97,21 +78,6 @@ class WordPressAuthenticatorTests: XCTestCase {
         let wpComSite = "testuser.wordpress.com"
         XCTAssert(WordPressAuthenticator.isWPComDomain(wpComSite))
     }
-
-
-    func testValidateSiteForSignin() {
-        let loginFields = LoginFields()
-
-        loginFields.siteAddress = ""
-        XCTAssert(!WordPressAuthenticator.validateSiteForSignin(loginFields), "Empty site should not validate.")
-
-        loginFields.siteAddress = "hostname"
-        XCTAssert(WordPressAuthenticator.validateSiteForSignin(loginFields), "Just a host name should validate.")
-
-        loginFields.siteAddress = "host name.com"
-        XCTAssert(!WordPressAuthenticator.validateSiteForSignin(loginFields), "Hostname with spaces should not validate.")
-    }
-
 
     func testEmailAddressTokenHandling() {
         let email = "example@email.com"
